+++ mercata/services/rewards-poller/.env.example
@@ -7,14 +7,18 @@ OPENID_DISCOVERY_URL=https://your-openid-provider/.well-known/openid-configurati
 
 # Contract Addresses
 REWARDS_CONTRACT_ADDRESS=0000000000000000000000000000000000000000
+PRICE_ORACLE_ADDRESS=0000000000000000000000000000000000000000
 USDST_ADDRESS=937efa7e3a77e20bbdbd7c0d32b6514f368c1010
 
 # API Configuration
 NODE_URL=https://your-strato-node-url
 
 # Polling Configuration
-POLLING_INTERVAL=60000
-MAX_BATCH_SIZE=100
+POLLING_INTERVAL=600000
+MAX_BATCH_SIZE=50
+
+# Bonus Configuration
+BONUS_CRON_SCHEDULE="0 3,9,15,21 * * *"
 
 # Balance Configuration
 # GAS_FEE_USDST: Gas fee in USDST (multiplied by 1e16, default: 1 = 0.01 USDST)
@@ -23,8 +27,8 @@ GAS_FEE_USDST=1
 GAS_FEE_VOUCHER=100
 # MIN_TRANSACTIONS_THRESHOLD: Minimum transactions that can be processed with current balance (default: 1)
 MIN_TRANSACTIONS_THRESHOLD=1
-# WARNING_TRANSACTIONS_THRESHOLD: Threshold for low balance warning (default: 50)
-WARNING_TRANSACTIONS_THRESHOLD=50
+# WARNING_TRANSACTIONS_THRESHOLD: Threshold for low balance warning (default: 100)
+WARNING_TRANSACTIONS_THRESHOLD=100
 
 # Retry Configuration
 RETRY_MAX_ATTEMPTS=2
