--- mercata/services/rewards-poller/src/config/bonusTokenConfig.json
@@ -1,3 +1,0 @@
-[
-  { "address": "00000000000000000000000000000000000000a1", "bonusBps": 500, "minBalance": "1" }
-]
