+++ mercata/services/rewards-poller/src/features/polling/polling.scheduler.ts
@@ -1,0 +1,15 @@
+import { logInfo } from "../../infra/observability/logger";
+
+export const startPollingLoop = (
+  component: string,
+  interval: number,
+  fn: () => Promise<void>
+): void => {
+  const poll = async () => {
+    await fn();
+    setTimeout(poll, interval);
+  };
+
+  void poll();
+  logInfo(component, `Started polling with interval ${interval}ms`);
+};
