+++ mercata/services/rewards-poller/src/infra/config/attributeMapping.json
