+++ mercata/services/rewards-poller/src/infra/config/bonusTokenConfig.json
@@ -1,0 +1,3 @@
+[
+  { "address": "3103810e081d965f5f8b19195fd7f5f859ccb4f9", "bonusBps": 1000, "minBalance": "1" }
+]
