+++ mercata/services/rewards-poller/src/infra/state/atomicJson.store.ts
@@ -1,0 +1,19 @@
+import { promises as fs } from "fs";
+import path from "path";
+
+export const writeJsonFileAtomic = async (
+  filePath: string,
+  data: unknown,
+): Promise<void> => {
+  const dir = path.dirname(filePath);
+  await fs.mkdir(dir, { recursive: true });
+
+  const tmpPath = `${filePath}.tmp.${process.pid}.${Date.now()}`;
+  try {
+    await fs.writeFile(tmpPath, JSON.stringify(data, null, 2), "utf-8");
+    await fs.rename(tmpPath, filePath);
+  } catch (error) {
+    await fs.unlink(tmpPath).catch(() => {});
+    throw error;
+  }
+};
