+++ mercata/services/rewards-poller/src/shared/core/address.ts
@@ -1,0 +1,2 @@
+export const normalizeAddressNoPrefix = (address: string): string =>
+  address.toLowerCase().replace(/^0x/, "");
