--- mercata/services/rewards-poller/src/utils/stratoHelper.ts
@@ -1,97 +1,0 @@
-import { bloc, strato, extractErrorMessage } from "./api";
-import { config } from "../config";
-import { logInfo } from "./logger";
-import { FunctionInput, BuiltTx, TxResult, TxResponse } from "../types";
-
-export const buildFunctionTx = (
-  inputs: FunctionInput | FunctionInput[],
-): BuiltTx => ({
-  txs: (Array.isArray(inputs) ? inputs : [inputs]).map(
-    ({ contractName, contractAddress, method, args }) => ({
-      type: config.strato.tx.type,
-      payload: { contractName, contractAddress, method, args },
-    }),
-  ),
-  txParams: {
-    gasLimit: config.strato.gas.limit,
-    gasPrice: config.strato.gas.price,
-  },
-});
-
-export const until = async <T>(
-  predicate: (result: T) => boolean,
-  action: () => Promise<T>,
-  {
-    timeout = config.strato.polling.defaultTimeout,
-    interval = config.strato.polling.defaultInterval,
-  }: { timeout?: number; interval?: number } = {},
-): Promise<T> => {
-  const deadline = Date.now() + timeout;
-
-  while (true) {
-    const result = await action();
-    if (predicate(result) || Date.now() >= deadline) return result;
-    await new Promise((r) => setTimeout(r, interval));
-  }
-};
-
-export const postAndWaitForTx = async (
-  postFn: () => Promise<any>,
-  timeout = config.strato.polling.defaultTimeout,
-): Promise<TxResponse> => {
-  const response = await postFn();
-  if (!Array.isArray(response) || !response.length) {
-    throw new Error("Invalid transaction response");
-  }
-
-  const txHashes = response.map((r, i) => {
-    if (!r?.hash) throw new Error(`Invalid tx result at index ${i}`);
-    return r.hash;
-  });
-
-  const results = await until(
-    (res: TxResult[]) => {
-      const failed = res.find((r) => r?.status === "Failure");
-      if (failed) {
-        const msg =
-          failed.txResult?.message ||
-          failed.error ||
-          failed.message ||
-          "Transaction failed";
-        throw new Error(extractErrorMessage(msg));
-      }
-      return res.every((r) => r?.status !== "Pending");
-    },
-    () => bloc.post("/transactions/results", txHashes),
-    { timeout },
-  );
-
-  return {
-    status: results[0].status as TxResponse["status"],
-    hash: results[0].hash,
-  };
-};
-
-export const execute = async (
-  inputs: FunctionInput | FunctionInput[],
-  timeout?: number,
-): Promise<TxResponse> => {
-  const inputArray = Array.isArray(inputs) ? inputs : [inputs];
-  const { method = "unknown", contractName = "unknown" } = inputArray[0] || {};
-  const context = `${method} on ${contractName}`;
-
-  logInfo("StratoHelper", `Executing ${context} (${inputArray.length} tx)`);
-
-  const result = await postAndWaitForTx(
-    () =>
-      strato.post(
-        "/transaction/parallel?resolve=true",
-        buildFunctionTx(inputs),
-      ),
-    timeout,
-  );
-
-  logInfo("StratoHelper", `${result.status}: ${context} (${result.hash})`);
-  return result;
-};
-
